# Conflicts:
#	app/src/main/java/ru/terrakok/gitlabclient/entity/Label.kt
#	app/src/main/java/ru/terrakok/gitlabclient/model/data/server/GitlabApi.kt
#	app/src/main/java/ru/terrakok/gitlabclient/model/interactor/project/ProjectInteractor.kt
#	app/src/main/java/ru/terrakok/gitlabclient/model/repository/project/ProjectRepository.kt

+++ app/src/main/java/ru/terrakok/gitlabclient/entity/Label.kt
@@ -2,8 +2,12 @@ package ru.terrakok.gitlabclient.entity
 
 import com.google.gson.annotations.SerializedName
 
+/**
+ * @author Maxim Myalkin (MaxMyalkin) on 29.10.2018.
+ */
+
 data class Label(
-    @SerializedName("id") val id: Int,
+    @SerializedName("id") val id: Long,
     @SerializedName("name") val name: String,
     @SerializedName("color") val color: String,
     @SerializedName("description") val description: String?,
